Bug: UTF-32-LE misdetected as UTF-16-LE
Problem: the byte-order mark detector selects the the UTF-16-LE
encoding if the first two bytes of the stream are OxFF OxFE, without
checking to see if the next two bytes are null (which is the BOM for
UTF-32-LE).

Solution: check for the UTF-32 BOMs before checking for the UTF-16
marks.